backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
aiosqlite==0.21.0
asyncpg==0.30.0
pydantic==2.11.1
pydantic-settings==2.8.1
PyJWT==2.10.1
stripe==11.6.0
google-genai==0.8.0
openai==1.59.5
anthropic==0.42.0
